+++ tests/beamlines/test_i19nxs.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from pathlib import Path
 from unittest.mock import patch
 
@@ -17,7 +19,7 @@ def test_nexus_writer_fails_if_missing_n_imgs_and_not_using_meta(fake_log):
         nexus_writer("/path/to/meta", "eiger", 0.1, use_meta=False)
 
 
-@patch("nexgen.beamlines.I19_2_nxs.tr_collect")
-def test_eiger_nxs_writer_fails_if_missing_axes_and_no_meta(fake_tr):
+@patch("nexgen.beamlines.I19_2_nxs.CollectionParams")
+def test_eiger_nxs_writer_fails_if_missing_axes_and_no_meta(fake_params):
     with pytest.raises(ValueError):
-        eiger_writer(Path("/path/to/meta"), fake_tr, use_meta=False)
+        eiger_writer(Path("/path/to/meta"), fake_params, use_meta=False)